+++ @kiva/kv-components/src/vue/stories/KvLoanUse.stories.js
@@ -52,3 +52,30 @@ export const WhySpecial = story({
 	name: 'Bob Smith',
 	whySpecial: 'It supports organic farming and includes a lower interest rate.',
 });
+
+export const BoldName = story({
+	use: 'buy supplies.',
+	loanAmount: '1000.00',
+	status: 'fundraising',
+	name: 'Bob Smith',
+	boldName: true,
+});
+
+export const BoldNameWithCountry = story({
+	use: 'buy supplies.',
+	loanAmount: '1000.00',
+	status: 'fundraising',
+	name: 'Bob Smith',
+	boldName: true,
+	country: 'Kenya',
+});
+
+export const IndicativeHelpText = story({
+	use: 'buy supplies.',
+	loanAmount: '1000.00',
+	status: 'fundraising',
+	name: 'Bob Smith',
+	hideLoanAmount: true,
+	useIndicativeHelpText: true,
+	whySpecial: 'It supports organic farming and includes a lower interest rate.',
+});