+++ @kiva/kv-components/src/vue/stories/KvUserAvatar.stories.js
@@ -11,9 +11,47 @@ const story = (args) => {
 		components: { KvUserAvatar },
 		setup() { return { args: templateArgs }; },
 		template: `
-			<KvUserAvatar v-bind="args" :is-small="false"/>
+			<KvUserAvatar v-bind="args" class="tw-w-6 tw-h-6"/>
 			<br/>
-			<KvUserAvatar v-bind="args" :is-small="true" />
+			<KvUserAvatar v-bind="args" class="tw-w-3 tw-h-3" is-small />
+		`,
+	});
+	template.args = args;
+	return template;
+};
+
+const cssPlaceholderStory = (args) => {
+	const template = (templateArgs, { argTypes }) => ({
+		props: Object.keys(argTypes),
+		components: { KvUserAvatar },
+		setup() {
+			const style = [];
+			if (args.lenderImageUrl) {
+				style.push(`--user-avatar: url(${args.lenderImageUrl}) / "Lender avatar via CSS"`);
+				style.push('--user-avatar-legacy-display: none');
+			} else {
+				style.push('--user-avatar-display: none');
+			}
+			return {
+				args: templateArgs,
+				style,
+			};
+		},
+		template: `
+			<div class="tw-relative" :style="style">
+				<KvUserAvatar
+					v-bind="args"
+					class="tw-w-6 tw-h-6"
+					show-css-placeholder
+				/>
+				<br/>
+				<KvUserAvatar
+					v-bind="args"
+					class="tw-w-3 tw-h-3"
+					is-small
+					show-css-placeholder
+				/>
+			</div>
 		`,
 	});
 	template.args = args;
@@ -21,8 +59,13 @@ const story = (args) => {
 };
 
 export const Default = story({
-	lenderImageUrl: 'https://www.development.kiva.org/img/s100/26e15431f51b540f31cd9f011cc54f31.jpg',
+	lenderImageUrl: 'https://www.kiva.org/img/s100/26e15431f51b540f31cd9f011cc54f31.jpg',
 	lenderName: 'Roger',
+});
+
+export const CssPlaceholder = cssPlaceholderStory({
+	lenderImageUrl: 'https://www.kiva.org/img/s100/26e15431f51b540f31cd9f011cc54f31.jpg',
+	showCssPlaceholder: true,
 });
 
 export const NoImage = story({
@@ -30,20 +73,19 @@ export const NoImage = story({
 	lenderName: 'Roger',
 });
 
+export const NoImageCssPlaceholder = cssPlaceholderStory({
+	lenderImageUrl: '',
+	showCssPlaceholder: true,
+});
+
 export const Anonymous = story({
-	lenderImageUrl: 'https://www.development.kiva.org/img/s100/26e15431f51b540f31cd9f011cc54f31.jpg',
+	lenderImageUrl: 'https://www.kiva.org/img/s100/26e15431f51b540f31cd9f011cc54f31.jpg',
 	lenderName: 'Anonymous',
 });
 
-export const DefaultProfile = story({
-	lenderImageUrl: 'https://www.development.kiva.org/img/s100/4d844ac2c0b77a8a522741b908ea5c32.jpg',
+export const LegacyDefaultImage = story({
+	lenderImageUrl: 'https://www.kiva.org/img/s100/4d844ac2c0b77a8a522741b908ea5c32.jpg',
 	lenderName: 'Default Profile',
-});
-
-export const IsSmall = story({
-	lenderImageUrl: 'https://www.development.kiva.org/img/s100/26e15431f51b540f31cd9f011cc54f31.jpg',
-	lenderName: 'Roger',
-	isSmall: true,
 });
 
 const multipleAvatarStory = (args) => {
@@ -58,6 +100,7 @@ const multipleAvatarStory = (args) => {
 					:key="index"
 					:lender-image-url="lender.lenderImageUrl"
 					:lender-name="lender.lenderName"
+					class="tw-w-6 tw-h-6"
 				/>
 			</div>
 		`,
@@ -69,15 +112,15 @@ const multipleAvatarStory = (args) => {
 export const Multiple = multipleAvatarStory({
 	lendersImages: [
 		{
-			lenderImageUrl: 'https://www.development.kiva.org/img/w500h500/0a35ae21956f8a03818b597877492196.jpg',
+			lenderImageUrl: 'https://www.kiva.org/img/w500h500/0a35ae21956f8a03818b597877492196.jpg',
 			lenderName: 'Emma',
 		},
 		{
-			lenderImageUrl: 'https://www.development.kiva.org/img/w500h500/957d6191b6f7b33f64fc49bab14bf108.jpg',
+			lenderImageUrl: 'https://www.kiva.org/img/w500h500/957d6191b6f7b33f64fc49bab14bf108.jpg',
 			lenderName: 'Oliver',
 		},
 		{
-			lenderImageUrl: 'https://www.development.kiva.org/img/w500h500/f0012a5353a2fe29d66364dc1d4e49ff.jpg',
+			lenderImageUrl: 'https://www.kiva.org/img/w500h500/f0012a5353a2fe29d66364dc1d4e49ff.jpg',
 			lenderName: 'John',
 		},
 	],